+++ DoW Mod Manager/DoW Mod Manager.csproj
@@ -7,7 +7,7 @@
     <ProjectGuid>{071DFF9E-1A7F-4F94-A1CE-4B5CB931D8EE}</ProjectGuid>
     <OutputType>WinExe</OutputType>
     <RootNamespace>DoW_Mod_Manager</RootNamespace>
-    <AssemblyName>DoW Mod Manager v2.4.1.1</AssemblyName>
+    <AssemblyName>DoW Mod Manager v2.4.2.0</AssemblyName>
     <TargetFrameworkVersion>v4.8</TargetFrameworkVersion>
     <FileAlignment>512</FileAlignment>
     <AutoGenerateBindingRedirects>true</AutoGenerateBindingRedirects>
@@ -74,6 +74,8 @@
     <Compile Include="FontsManagerForm.Designer.cs">
       <DependentUpon>FontsManagerForm.cs</DependentUpon>
     </Compile>
+    <Compile Include="LAATweaks.cs" />
+    <Compile Include="LAATweaksOld.cs" />
     <Compile Include="MessageForm.cs">
       <SubType>Form</SubType>
     </Compile>
