+++ DoW Mod Manager/LAATweaksOld.cs
@@ -1,0 +1,98 @@
+﻿using System.IO;
+
+namespace DoW_Mod_Manager
+{
+    class LAATweaksOld
+    {
+        const int IMAGE_FILE_LARGE_ADDRESS_AWARE = 0x20;
+
+        /// <summary>
+        /// This method instigates the test if a given EXE is LAA patched or not.
+        /// </summary>
+        /// <param name="file"></param>
+        /// <returns>bool</returns>
+        public static bool IsLargeAware(string file)
+        {
+            using (FileStream fs = File.OpenRead(file))
+            {
+                using (BinaryReader br = new BinaryReader(fs))
+                {
+                    if (br.ReadInt16() != 0x5A4D)       // No MZ Header
+                        return false;
+
+                    br.BaseStream.Position = 0x3C;
+                    int peloc = br.ReadInt32();         // Get the PE header location.
+
+                    br.BaseStream.Position = peloc;
+                    if (br.ReadInt32() != 0x4550)       // No PE header
+                        return false;
+
+                    br.BaseStream.Position += 0x12;     // LAA flag position
+                    short LAAFlag = br.ReadInt16();
+
+                    return (LAAFlag & IMAGE_FILE_LARGE_ADDRESS_AWARE) == IMAGE_FILE_LARGE_ADDRESS_AWARE;
+                }
+            }
+        }
+
+        /// <summary>
+        /// This method performs the necessary data operations in order to toggle the LAA for a given EXE file back and forth.
+        /// </summary>
+        /// <param name="file"></param>
+        /// <returns>bool</returns>
+        public static bool ToggleLAA(string file)
+        {
+            bool result = false;
+
+            using (FileStream fs = File.Open(file, FileMode.Open, FileAccess.ReadWrite, FileShare.None))
+            {
+                BinaryReader br = new BinaryReader(fs);
+
+                if (br.ReadInt16() != 0x5A4D)       // No MZ Header
+                    return false;
+
+                br.BaseStream.Position = 0x3C;
+                int peloc = br.ReadInt32();         // Get the PE header location.
+
+                br.BaseStream.Position = peloc;
+                if (br.ReadInt32() != 0x4550)       // No PE header
+                    return false;
+
+                br.BaseStream.Position += 0x12;     // LAA flag position
+                long nFilePos = (int)br.BaseStream.Position;
+                short LAAFlag = br.ReadInt16();
+
+                br.BaseStream.Position += 0x40;     // Checksum position
+                long nSumPos = (int)br.BaseStream.Position;
+                short ChckSum = br.ReadInt16();
+
+                BinaryWriter bw = new BinaryWriter(fs);
+                if ((LAAFlag & IMAGE_FILE_LARGE_ADDRESS_AWARE) != IMAGE_FILE_LARGE_ADDRESS_AWARE)
+                {
+                    LAAFlag += IMAGE_FILE_LARGE_ADDRESS_AWARE;
+                    ChckSum += IMAGE_FILE_LARGE_ADDRESS_AWARE;
+                    bw.Seek((int)nFilePos, SeekOrigin.Begin);
+                    bw.Write(LAAFlag);
+                    bw.Seek((int)nSumPos, SeekOrigin.Begin);
+                    bw.Write(ChckSum);
+                    bw.Flush();
+                    result = true;
+                }
+                else if ((LAAFlag & IMAGE_FILE_LARGE_ADDRESS_AWARE) == IMAGE_FILE_LARGE_ADDRESS_AWARE)
+                {
+                    LAAFlag -= IMAGE_FILE_LARGE_ADDRESS_AWARE;
+                    ChckSum -= IMAGE_FILE_LARGE_ADDRESS_AWARE;
+                    bw.Seek((int)nFilePos, SeekOrigin.Begin);
+                    bw.Write(LAAFlag);
+                    bw.Seek((int)nSumPos, SeekOrigin.Begin);
+                    bw.Write(ChckSum);
+                    bw.Flush();
+                    result = false;
+                }
+                bw.Close();
+                br.Close();
+            }
+            return result;
+        }
+    }
+}
