+++ DoW Mod Manager/LatestStable/latestchangelog
@@ -1,0 +1,5 @@
+## Version 2.4.2.0
+
+- Added: changelog in the Application Update and Mod Update dialog. Now users could see what exactly has changed in the last version!
+- Fixed: Applying LAA patch to Soulstorm.exe could cause Multiplayer Lobby List to be always empty. To fix that, user should disable LAA, but after Mod Manager version 2.4.0.0 that would not fix the issue! That bug has been fixed in this update! GOG version executable would be patched using new method while Soulstorm executable would be patched using old, simplistic but more reliable method.
+- Refactored: moved all LAA patch logic to another class
