+++ DoW Mod Manager/LatestStable/version
@@ -1,1 +1,1 @@
-2.4.1.1
+2.4.2.0
