+++ DoW Mod Manager/ModList/latestchangelog
@@ -1,0 +1,1 @@
+Updated Soulstorm Performance mod to version 1.3.0
