+++ changelog.txt
@@ -2,8 +2,14 @@
 
 Who did what:
 Fragjacker (Original author) - up to 1.5.2,  then 1.6.1,  2.0.3,  2.0.5,  2.1.2,  2.1.4,  2.1.5,  2.1.7,  2.1.9,,  2.2.2,  2.3.1.0
-IgorTheLight (current main maintainer) - 1.5.3 to 1.6,  1.7 to 2.0.3,  2.0.4,  2.1.0,  2.1.1,  2.1.3,  2.1.6,  2.1.8,  2.2.0,  2.2.1,  2.3.0.0, 2.3.2.0, 2.4.0.0, 2.4.1.0, 2.4.1.1
+IgorTheLight (current main maintainer) - 1.5.3 to 1.6,  1.7 to 2.0.3,  2.0.4,  2.1.0,  2.1.1,  2.1.3,  2.1.6,  2.1.8,  2.2.0,  2.2.1,  2.3.0.0, 2.3.2.0, 2.4.0.0, 2.4.1.0, 2.4.1.1, 2.4.2.0
 BlueAmulet - 2.1.7
+
+## Version 2.4.2.0
+
+- Added: changelog in the Application Update and Mod Update dialog. Now users could see what exactly has changed in the last version!
+- Fixed: Applying LAA patch to Soulstorm.exe could cause Multiplayer Lobby List to be always empty. To fix that, user should disable LAA, but after Mod Manager version 2.4.0.0 that would not fix the issue! That bug has been fixed in this update! GOG version executable would be patched using new method while Soulstorm executable would be patched using old, simplistic but more reliable method.
+- Refactored: moved all LAA patch logic to another class
 
 ## Version 2.4.1.1
